200,201,404

+++ core/views.py
@@ -34,7 +34,7 @@ def note_detail(request, pk):
   except Note.DoesNotExist:
       return Response({"detail": "Not found."}, status=status.HTTP_404_NOT_FOUND)
   
-  if request.method =="Get":
+  if request.method =="GET":
      serializer = NoteSerializer(note)
      return Response(serializer.data)
   
@@ -43,7 +43,8 @@ def note_detail(request, pk):
      if serializer.is_valid():
         serializer.save()
         return Response(serializer.data)
-     return Response(serializer.erros, status.HTTP_400_BAD_REQUEST)
+     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
   
-  note.delete()
-  return Response(status=status.HTTP_204_NO_CONTENT)
+  elif request.method =="DELETE":
+    note.delete()
+    return Response(status=status.HTTP_204_NO_CONTENT)
